requirements.txt
fastapi==0.115.2
torch==2.5.0
torchvision==0.20.0
uvicorn==0.32.0
requests==2.32.3
python-multipart==0.0.12
faiss-cpu==1.9.0
Pillow==11.0.0
numpy==2.1.2
httpx==0.27.2
pytest==8.3.3
